+++ router-app/src/Components/Dashboard/NotesList.tsx
@@ -1,64 +1,29 @@
 import { useEffect } from "react";
 import useFetch from "../../hooks/useFetch";
-import { useNavigate } from "react-router-dom";
-import { category } from "../../types";
+import { category, NoteCompact } from "../../types";
+import NotesTable from "./NotesTable";
 
 type NotesListProps = {
     selectedCategory: category | null;
 }
 
-type noteCompact = {
-    guid: string;
-    title: string;
-    categoryGuid: string;
-    categoryName: string;
-    dateAdded: string;
-    lastChanged: string;
-}
-
 export default function NotesList({ selectedCategory }: NotesListProps) {
-    const { data, isPending, error, doFetch } = useFetch<noteCompact[]>("/GetNotesByCategory", []);
+    const { data, isPending, error, doFetch } = useFetch<NoteCompact[]>("/GetNotesByCategory", []);
 
     useEffect(() => {
         doFetch("GET", [selectedCategory?.guid || ""]);
     }, [selectedCategory])
 
-    const navigate = useNavigate();
+    const headerText = selectedCategory !== null ? `${data?.length} Note(s) with category "${selectedCategory.name}"` : `${data?.length} Uncategorized note(s)`;
 
     return (
         <>
-            <h1>{selectedCategory !== null ? `${data?.length} Note(s) with category "${selectedCategory.name}"` : `${data?.length} Uncategorized note(s)`}</h1>
-            {isPending && <span>Loading notes...</span>}
-            {error && <span>{error}</span>}
-            {
-                data && !error && !isPending &&
-                <table>
-                    <tbody>
-                        <tr>
-                            <th>Title</th>
-                            <th>Category</th>
-                            <th>Last Changed</th>
-                            <th>Date Added</th>
-                        </tr>
-                        {data?.map((note) => (
-                            <tr key={note.dateAdded} onClick={() => navigate("/note/" + note.guid)}>
-                                <td>
-                                    {note.title}
-                                </td>
-                                <td>
-                                    {note.categoryName}
-                                </td>
-                                <td>
-                                    {note.lastChanged}
-                                </td>
-                                <td>
-                                    {note.dateAdded}
-                                </td>
-                            </tr>
-                        ))}
-                    </tbody>
-                </table>
-            }
+            <h1>
+                {data && headerText}
+                {isPending && <>Loading notes...</>}
+                {error && <>{error}</>}
+            </h1>
+            {data && <NotesTable data={data} />}
         </>
     )
 }
